# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: EduEarn Student Productivity & Earning API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tutors:
    get:
      operationId: getTutors
      tags: [tutors]
      summary: Get tutor/coaching listings
      parameters:
        - name: city
          in: query
          schema:
            type: string
        - name: subject
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of tutors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TutorListing"
    post:
      operationId: createTutorListing
      tags: [tutors]
      summary: Create a tutor listing (teacher role only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTutorListingRequest"
      responses:
        "201":
          description: Listing created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TutorListing"
        "403":
          description: Teacher role required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /tutors/{id}:
    delete:
      operationId: deleteTutorListing
      tags: [tutors]
      summary: Delete own tutor listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
...
    TutorListing:
      type: object
      properties:
        id:
          type: integer
        teacherId:
          type: integer
        teacherName:
          type: string
        subject:
          type: string
        description:
          type: string
          nullable: true
        fees:
          type: integer
        feesLabel:
          type: string
        city:
          type: string
        contactEmail:
          type: string
        contactPhone:
          type: string
          nullable: true
        mode:
          type: string
          enum: [online, offline, both]
        createdAt:
          type: string
          format: date-time
      required: [id, teacherId, teacherName, subject, fees, feesLabel, city, contactEmail, mode, createdAt]
    CreateTutorListingRequest:
      type: object
      properties:
        subject:
          type: string
        description:
          type: string
        fees:
          type: integer
        feesLabel:
          type: string
        city:
          type: string
        contactEmail:
          type: string
        contactPhone:
          type: string
        mode:
          type: string
          enum: [online, offline, both]
      required: [subject, fees, feesLabel, city, contactEmail, mode]
...